moved factorial to own file
